Add Powerups
Enemies may be given powerups as part of the level definition.
When an enemy carrying a powerup is destroyed, it drops the
powerup, which travels on the enemy's last trajectory with the
same velocity as the enemy.  When the player touches a powerup,
the powerup is added to the player.  When a player is damaged, he
loses a powerup at random.