Enemy cannons pivot to follow player
The enemy cannons will rotate to point at the player within a
certain pivot range.  This causes the cannon drones to target the
player when the player is within their arc of fire.

During this update I realized that part of the reason some of the
angle math seems so broken is that the Love graphics origin is in
the upper-left corner of the display, and positive values of Y
move down the screen.  This causes a lot of the trigonometry
stuff to seem weird, since it basically makes all angles
negative.  I'm not sure if this is worth fixing for this project,
but would be handy to keep in mind for the future.