Make transition screen state, add level complete
Instead of having states for every text screen make a single
transition screen that handles all of them.  Reduces code and
duplication and makes it easier to modify behavior.

This also adds a level complete screen as another transition
screen and causes the "win" and "gameover" screens to loop back
to the title.